out: Elfx32, Elf32 -- Unify elf_add_reloc
Make offset being int64_t and use 0 for Elf32.
Thus it gonna be the same for all Elfs.

+++ output/outelfx32.c
@@ -557,7 +557,7 @@ static void elf_deflabel(char *name, int32_t segment, int64_t offset,
 }
 
 static void elf_add_reloc(struct elf_section *sect, int32_t segment,
-                          int32_t offset, int type)
+                          int64_t offset, int type)
 {
     struct elf_reloc *r;
